Exception handling for user retrieval

+++ example-microservice/src/main/java/nl/tudelft/sem/template/example/controllers/UsersController.java
@@ -58,6 +58,42 @@ public class UsersController {
     AccountSettingsController accountSettingsController;
     UserDetailsController userDetailsController;
 
+    /**
+     * Utility function for the checking below.
+     * @param field A list of objects to go through and check for nulls
+     * @return whether the list contains nulls or not
+     */
+    private boolean arrayHasNullsUtility(List<?> field) {
+        for (Object o : field)
+            if (o == null)
+                return true;
+        return false;
+    }
+
+    /**
+     * Utility function for error checking when trying to retrieve a user from the database.
+     * @param userID the ID of the user to retrieve
+     * @param message1 the first error message, corresponding to the first error status code
+     * @param message2 the second error message, corresponding to the second error status code
+     * @return ResponseEntity of either a String with the error message, or the actual User retrieved
+     */
+    private ResponseEntity<? extends Object> existenceCheckingWithCustomMessages(Integer userID, String message1, String message2) {
+        User user;
+        try {
+            Optional<User> optionalUser = userRepository.findById(userID);
+            if(optionalUser.isEmpty())
+                throw new NoSuchElementException();
+            user = optionalUser.get();
+        }
+        catch (NoSuchElementException e) {
+            return new ResponseEntity<>(message1, HttpStatus.NOT_FOUND);
+        }
+        catch (Exception e) {
+            return new ResponseEntity<>(message2, HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+        return new ResponseEntity<>(user, HttpStatus.OK);
+    }
+
     @Autowired
     public UsersController(UserRepository userRepository,
                            UserRegistrationService userRegistrationService,
@@ -210,18 +246,10 @@ public class UsersController {
     )
     public ResponseEntity<String> makeAdmin(@PathVariable(name = "userID") int userId,
                                             @RequestBody String password) {
-        User user;
-        try {
-            Optional<User> optionalUser = userRepository.findById(userId);
-            if (optionalUser.isEmpty()) {
-                throw new NoSuchElementException();
-            }
-            user = optionalUser.get();
-        } catch (NoSuchElementException e) {
-            return new ResponseEntity<>("Username with that ID could not be found", HttpStatus.NOT_FOUND);
-        } catch (Exception e) {
-            return new ResponseEntity<>("Something went wrong", HttpStatus.INTERNAL_SERVER_ERROR);
-        }
+        ResponseEntity<?> checkUserIDResult = existenceCheckingWithCustomMessages(userId, "User with that ID could not be found", "Something went wrong");
+        if (checkUserIDResult.getBody() instanceof String)
+            return new ResponseEntity<>((String)checkUserIDResult.getBody(), checkUserIDResult.getStatusCode());
+        User user = (User) checkUserIDResult.getBody();
 
         try {
             verificationService.verifyAdminRequest(user, password);
@@ -231,89 +259,15 @@ public class UsersController {
             return new ResponseEntity<>(e.getMessage(), HttpStatus.CONFLICT);
         }
 
-        String message = "User with ID:" + userId + " is now an admin";
         user.setIsAdmin(true);
         try {
             userRepository.save(user);
         } catch (Exception e) {
             return new ResponseEntity<>("User could not be updated", HttpStatus.INTERNAL_SERVER_ERROR);
         }
+
+        String message = "User with ID:" + userId + " is now an admin";
         return new ResponseEntity<>(message, HttpStatus.OK);
-    }
-
-    /**
-     * GET /user/{userID} : Fetch user information.
-     *
-     * @param userID Numeric ID of the user that makes the request (required)
-     * @return User data fetched successfully (status code 200)
-     *         or User is not authenticated (status code 401)
-     *         or Internal server error (status code 500)
-     */
-    @GetMapping(value = "/user/{userID}")
-    public ResponseEntity<User> userGetUser(
-            @Parameter(name = "userID", description = "Numeric ID of the user that makes the request", required = true)
-            @PathVariable("userID") Integer userID) {
-
-        User user;
-        try {
-            user = userRegistrationService.getUserById(userID);
-        } catch (Exception e) {
-            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
-        }
-
-        if (user != null) {
-            return new ResponseEntity<>(user, HttpStatus.OK);
-        } else {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-        }
-    }
-
-    /**
-     * PUT /user/{userID}/changePassword : Change user password.
-     *
-     * @param userId Numeric ID of the user that makes the request (required)
-     * @param body  The desired password (required)
-     * @return Password changed successfully (status code 200)
-     *         or Request body is malformed (status code 401)
-     *         or Password could not be changed (status code 500)
-     */
-    @PutMapping(value = "/user/{userID}/changePassword")
-    public ResponseEntity<String> userChangePassword(
-            @Parameter(name = "userID", required = true) @PathVariable("userID") Integer userId,
-            @Parameter(name = "body", description = "Desired Password", required = true) @Valid @RequestBody String body
-    ) {
-        if (userId == null || body.isEmpty()) {
-            return new ResponseEntity<>("Request body is malformed", HttpStatus.BAD_REQUEST);
-        }
-
-        User user;
-        try {
-            Optional<User> optionalUser = userRepository.findById(userId);
-            if (optionalUser.isEmpty()) {
-                throw new NoSuchElementException();
-            }
-            user = optionalUser.get();
-        } catch (NoSuchElementException e) {
-            return new ResponseEntity<>("User with that ID could not be found", HttpStatus.NOT_FOUND);
-        } catch (IllegalArgumentException e) {
-            return new ResponseEntity<>("Something went wrong", HttpStatus.INTERNAL_SERVER_ERROR);
-        }
-
-        HashedPassword hashedPassword = PasswordHashingService.hash(body);
-        User updatedUser;
-
-        try {
-            updatedUser = updateUserService.changePassword(user.getId(), hashedPassword);
-            if (updatedUser == null) {
-                throw new NoSuchElementException();
-            }
-        } catch (NoSuchElementException e) {
-            return new ResponseEntity<>("Couldn't change the password", HttpStatus.NOT_FOUND);
-        } catch (Exception e) {
-            return new ResponseEntity<>("Database insertion failed", HttpStatus.INTERNAL_SERVER_ERROR);
-        }
-
-        return new ResponseEntity<>("Your password has been changed successfully.", HttpStatus.OK);
     }
 
     /**
@@ -340,18 +294,10 @@ public class UsersController {
         }
         int documentId = document.getDocumentID();
 
-        User user;
-        try {
-            Optional<User> optionalUser = userRepository.findById(userId);
-            if (optionalUser.isEmpty()) {
-                throw new NoSuchElementException();
-            }
-            user = optionalUser.get();
-        } catch (NoSuchElementException e) {
-            return new ResponseEntity<>("User with that ID could not be found", HttpStatus.NOT_FOUND);
-        } catch (Exception e) {
-            return new ResponseEntity<>("Something went wrong", HttpStatus.INTERNAL_SERVER_ERROR);
-        }
+        ResponseEntity<?> checkUserIDResult = existenceCheckingWithCustomMessages(userId, "User with that ID could not be found", "Something went wrong");
+        if (checkUserIDResult.getBody() instanceof String)
+            return new ResponseEntity<>((String)checkUserIDResult.getBody(), checkUserIDResult.getStatusCode());
+        User user = (User) checkUserIDResult.getBody();
 
         try {
             verificationService.verifyAuthorRequest(user, documentId);
@@ -361,14 +307,79 @@ public class UsersController {
             return new ResponseEntity<>(e.getMessage(), HttpStatus.UNAUTHORIZED);
         }
 
-        String message = "User with ID:" + userId + " is now an author";
         user.setIsAuthor(true);
         try {
             userRepository.save(user);
-        } catch (Exception e) {
+        }
+        catch (Exception e) {
             return new ResponseEntity<>("User could not be saved", HttpStatus.INTERNAL_SERVER_ERROR);
         }
+
+        String message = "User with ID:" + userId + " is now an author";
         return new ResponseEntity<>(message, HttpStatus.OK);
+    }
+
+    /**
+     * GET /user/{userID} : Fetch user information.
+     *
+     * @param userID Numeric ID of the user that makes the request (required)
+     * @return User data fetched successfully (status code 200)
+     *         or User is not authenticated (status code 401)
+     *         or Internal server error (status code 500)
+     */
+    @GetMapping(value = "/user/{userID}")
+    public ResponseEntity<User> userGetUser(
+            @Parameter(name = "userID", description = "Numeric ID of the user that makes the request", required = true)
+            @PathVariable("userID") Integer userID) {
+
+        User user;
+        try {
+            user = userRegistrationService.getUserById(userID);
+        } catch (Exception e) {
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+
+        return user != null ? new ResponseEntity<>(user, HttpStatus.OK) : new ResponseEntity<>(HttpStatus.NOT_FOUND);
+    }
+
+    /**
+     * PUT /user/{userID}/changePassword : Change user password.
+     *
+     * @param userId Numeric ID of the user that makes the request (required)
+     * @param body  The desired password (required)
+     * @return Password changed successfully (status code 200)
+     *         or Request body is malformed (status code 401)
+     *         or Password could not be changed (status code 500)
+     */
+    @PutMapping(value = "/user/{userID}/changePassword")
+    public ResponseEntity<String> userChangePassword(
+            @Parameter(name = "userID", required = true) @PathVariable("userID") Integer userId,
+            @Parameter(name = "body", description = "Desired Password", required = true) @Valid @RequestBody String body
+    ) {
+        if (userId == null || body.isEmpty()) {
+            return new ResponseEntity<>("Request body is malformed", HttpStatus.BAD_REQUEST);
+        }
+
+        ResponseEntity<?> checkUserIDResult = existenceCheckingWithCustomMessages(userId, "User with that ID could not be found", "Something went wrong");
+        if (checkUserIDResult.getBody() instanceof String)
+            return new ResponseEntity<>((String)checkUserIDResult.getBody(), checkUserIDResult.getStatusCode());
+        User user = (User) checkUserIDResult.getBody();
+
+        HashedPassword hashedPassword = PasswordHashingService.hash(body);
+        User updatedUser;
+
+        try {
+            updatedUser = updateUserService.changePassword(user.getId(), hashedPassword);
+            if (updatedUser == null) {
+                throw new NoSuchElementException();
+            }
+        } catch (NoSuchElementException e) {
+            return new ResponseEntity<>("Couldn't change the password", HttpStatus.NOT_FOUND);
+        } catch (Exception e) {
+            return new ResponseEntity<>("Database insertion failed", HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+
+        return new ResponseEntity<>("Your password has been changed successfully.", HttpStatus.OK);
     }
 
     /**
@@ -398,21 +409,15 @@ public class UsersController {
     public ResponseEntity<Void> userUserIDDelete(
             @Parameter(name = "userID", description = "Numeric ID of the user that makes the request", required = true, in = ParameterIn.PATH) @PathVariable("userID") Integer userID
     ) {
-        if(userID == null){
+        if(userID == null) {
             return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
         }
         //given this api, there is no way to verify whether an user is logged in or not ...
         //should we modify the api in order to also include in the url the id of the user making the request???
-        User user;
-        try {
-            Optional<User> optionalUser = userRepository.findById(userID);
-            if (optionalUser.isEmpty()) {
-                throw new NoSuchElementException();
-            }
-            user = optionalUser.get();
-        } catch (NoSuchElementException e) {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-        }
+        ResponseEntity<?> checkUserIDResult = existenceCheckingWithCustomMessages(userID, "User with that ID could not be found", "Something went wrong");
+        if (checkUserIDResult.getBody() instanceof String)
+            return new ResponseEntity<>(checkUserIDResult.getStatusCode());
+        User user = (User) checkUserIDResult.getBody();
         try {
             userRepository.delete(user);
         } catch (Exception e) {
@@ -441,13 +446,13 @@ public class UsersController {
                                                         @RequestParam List<String> interests) {
         if(interests == null)
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
-        for(String s : interests)
-            if(s == null)
+        if (arrayHasNullsUtility(interests))
                 return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
 
-        User user = userRepository.findById(userID).orElse(null);
-        if(user == null)
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        ResponseEntity<?> checkUserIDResult = existenceCheckingWithCustomMessages(userID, "User with that ID could not be found", "Something went wrong");
+        if (checkUserIDResult.getBody() instanceof String)
+            return new ResponseEntity<>(checkUserIDResult.getStatusCode());
+        User user = (User) checkUserIDResult.getBody();
         List<User> allUsers;
         try {
             allUsers = userRepository.findAll();
@@ -493,13 +498,13 @@ public class UsersController {
                                                         @RequestBody List<Book> favoriteBooks) {
         if (favoriteBooks == null)
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
-        for(Book b : favoriteBooks)
-            if(b == null)
+        if (arrayHasNullsUtility(favoriteBooks))
                 return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
 
-        User user = userRepository.findById(userID).orElse(null);
-        if(user == null)
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        ResponseEntity<?> checkUserIDResult = existenceCheckingWithCustomMessages(userID, "User with that ID could not be found", "Something went wrong");
+        if (checkUserIDResult.getBody() instanceof String)
+            return new ResponseEntity<>(checkUserIDResult.getStatusCode());
+        User user = (User) checkUserIDResult.getBody();
         List<Integer> ids = favoriteBooks.stream().map(Book::getId).collect(Collectors.toList());
         List<User> allUsers;
         try {
@@ -548,17 +553,15 @@ public class UsersController {
                                                               @RequestBody List<UserSearch> connections) {
         if(connections == null)
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
-        for(UserSearch u : connections)
-            if(u == null)
+        if (arrayHasNullsUtility(connections))
                 return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
-        if(userRepository.findById(userID).isEmpty())
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        ResponseEntity<?> checkUserIDResult = existenceCheckingWithCustomMessages(userID, "User with that ID could not be found", "Something went wrong");
+        if (checkUserIDResult.getBody() instanceof String)
+            return new ResponseEntity<>(checkUserIDResult.getStatusCode());
         List<Email> emails = connections.stream().map(UserSearch::getEmail).collect(Collectors.toList());
         // if any email does not match the correct format => Bad request
-        for(Email e : emails) {
-            if(e == null)
+        if (arrayHasNullsUtility(emails))
                 return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
-        }
         List<User> allUsers;
         try {
             allUsers = userRepository.findAll();
@@ -595,28 +598,22 @@ public class UsersController {
             @Parameter(name = "name", description = "Name of the searched user", required = true)
             @PathVariable String name) {
 
-
         if (name == null || name.isEmpty() || userId == null) {
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
 
-        try {
-            Optional<User> optionalUser = userRepository.findById(userId);
-            if (optionalUser.isEmpty()) {
-                throw new NoSuchElementException();
-            }
-        } catch (NoSuchElementException e) {
+        ResponseEntity<?> checkUserIDResult = existenceCheckingWithCustomMessages(userId, "User with that ID could not be found", "Something went wrong");
+        if (checkUserIDResult.getBody() instanceof String)
+            return new ResponseEntity<>(checkUserIDResult.getStatusCode());
+
+        List<User> users;
+        users = userRegistrationService.getUserByUsername(name);
+
+        if (!users.isEmpty()) {
+            return new ResponseEntity<>(users, HttpStatus.OK);
+        } else {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-
         }
-            List<User> users;
-            users = userRegistrationService.getUserByUsername(name);
-
-            if (!users.isEmpty()) {
-                return new ResponseEntity<>(users, HttpStatus.OK);
-            } else {
-                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-            }
         }
 
     /**
@@ -640,16 +637,9 @@ public class UsersController {
         // basic checking beforehand
         if(userID == null || anyID == null)
             return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
-        User user;
-        try {
-            Optional<User> optionalUser = userRepository.findById(userID);
-            if(optionalUser.isEmpty())
-                throw new InvalidUserException();
-            user = optionalUser.get();
-        }
-        catch (InvalidUserException e) {
-            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
-        }
+        ResponseEntity<?> checkUserIDResult = existenceCheckingWithCustomMessages(userID, "User with that ID could not be found", "Something went wrong");
+        if (checkUserIDResult.getBody() instanceof String)
+            return new ResponseEntity<>((String)checkUserIDResult.getBody(), checkUserIDResult.getStatusCode());
 
         Optional<? extends Object> optional = userDetailsRegistrationService.findById(anyID); // first look whether it was a UserDetails request
         if (optional.isEmpty()) {
